Move to funcs.
